morebits: Add undeletePage to morebits.wiki.page
Basically just copy the stuff for deletePage, less following redirects.  deletePage was added in 4b92869 along with move, but unlike protect delete and undelete are one-way streets.  Uses a dumb kludge to get around the lack of a proper token until we can rewrite the token handling stuff in `morebits`.

https://github.com/azatoth/twinkle/commit/3b0f12ea7771fd8ff3d82480de2a57a081d029ad